portfolio-backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic[email]==2.5.3
python-multipart==0.0.6
supabase==2.3.4
PyJWT==2.8.0
